changes property crud

+++ src/main/java/com/immovable/propertyservice/entities/Property.java
@@ -1,5 +1,6 @@
 package com.immovable.propertyservice.entities;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.annotation.Id;
@@ -24,9 +25,9 @@ public class Property {
 	private String imageUrl;
 	private String currency;
 	private String occupancyRate;
-	private Double actualpropertyPrice;
-	private Double transcationCostPrice;
-    private Double totalInvestmentCost;//  (actualpropertyPrice+transcationCostPrice)
+	private BigDecimal actualpropertyPrice;
+	private BigDecimal transcationCostPrice;
+    private BigDecimal totalInvestmentCost;//  (actualpropertyPrice+transcationCostPrice)
 	private String status;
 	@DBRef
 	private PropertyMetaData propertyMetaData;
@@ -34,6 +35,7 @@ public class Property {
 	private List<PropertyStakeInfo> propertyStakeInfo;
 	@DBRef
 	private List<PropertyRevenueInfo> propertyRevenueInfo;
+	
 	
 
 }
